flask/requirements.txt
Flask==1.0.2
PassportEye==1.4.0
pillow==6.0.0
pytesseract==0.2.6
opencv-python==4.1.0.25
